project-3-commod-etl/requirements.txt
pandas
SQLAlchemy
python-dotenv
requests
